+++ rxnbench/backend/client/src/rxn_bench_client/__init__.py
@@ -1,5 +1,13 @@
 from .client import RxnBenchClient, ExperimentStopped
 from .data_logger import DataLogger
-from .instruments import Camera, Gantry, PHProbe
+from .instruments import Camera, Gantry, PHProbe, PHStabilityTimeout
 
-__all__ = ["RxnBenchClient", "ExperimentStopped", "DataLogger", "Camera", "Gantry", "PHProbe"]
+__all__ = [
+    "RxnBenchClient",
+    "ExperimentStopped",
+    "DataLogger",
+    "Camera",
+    "Gantry",
+    "PHProbe",
+    "PHStabilityTimeout",
+]